refactor: create one source of truth for navItems
- move NavItem type to NavItems component
- make NavItems use the data from data/navItems
