+++ polaris.shopify.com/src/components/Layout/Layout.tsx
@@ -3,7 +3,9 @@ import { TOCItem, useTOC } from "../../utils/hooks";
 import { className, slugify } from "../../utils/various";
 import Longform from "../Longform";
 import Container from "../Container";
-import Nav, { NavItem } from "../Nav/Nav";
+import Nav from "../Nav";
+import type { NavItem } from "../NavItems";
+
 import styles from "./Layout.module.scss";
 
 interface Props {
