+++ polaris.shopify.com/src/components/Nav/Nav.tsx
@@ -1,12 +1,9 @@
 import Link from "next/link";
 import { useRouter } from "next/router";
-import styles from "./Nav.module.scss";
 
-export type NavItem = {
-  title: string;
-  url?: string;
-  children?: NavItem[];
-};
+import type { NavItem } from "../NavItems";
+
+import styles from "./Nav.module.scss";
 
 interface Props {
   navItems: NavItem[];
