+++ polaris.shopify.com/src/components/NavItems/index.ts
@@ -1,3 +1,4 @@
 import NavItems from "./NavItems";
+export type { NavItem } from "./NavItems";
 
 export default NavItems;
